Move to spiral.optimizer directory
