+++ megatron/optimizer/__init__.py
@@ -12,7 +12,7 @@ from megatron.spiral.utils import is_spiral_param
 from .distrib_optimizer import DistributedOptimizer
 from .grad_scaler import ConstantGradScaler, DynamicGradScaler
 from .optimizer import Float16OptimizerWithFloat16Params, FP32Optimizer
-from megatron.spiral.optimizer import SpiralStageOptimizer
+from megatron.spiral.optimizer.stage_optimizer import SpiralStageOptimizer
 
 
 def get_param_groups(modules,
@@ -147,7 +147,7 @@ def get_megatron_optimizer(model,
 
         # NOTE (SpiralPipe) Spiral stage optimizer uses SpiralCPUAdam, which overlaps weight update with upstream bwd stage computation.
         if args.spiral_stage_optimizer:
-            from megatron.spiral.cpu_adam import SpiralCPUAdam
+            from megatron.spiral.optimizer.cpu_adam import SpiralCPUAdam
             inner_opt_ty = SpiralCPUAdam
         else:
             from deepspeed.ops.adam import DeepSpeedCPUAdam
