+++ megatron/spiral/optimizer/stage_optimizer.py
